Lagt till Besluts statistik mm
